refactor: 소셜 로그인 추상화
