+++ src/main/kotlin/com/wq/auth/api/domain/auth/GoogleLoginProvider.kt
@@ -2,10 +2,11 @@ package com.wq.auth.api.domain.auth
 
 import com.wq.auth.api.domain.auth.entity.ProviderType
 import com.wq.auth.api.domain.auth.entity.RefreshTokenEntity
+import com.wq.auth.api.domain.auth.request.OAuthAuthCodeRequest
+import com.wq.auth.api.domain.auth.request.SocialLoginRequest
+import com.wq.auth.api.domain.auth.response.SocialLoginResult
 import com.wq.auth.api.domain.member.MemberRepository
 import com.wq.auth.api.external.oauth.GoogleOAuthClient
-import com.wq.auth.domain.auth.request.SocialLoginRequest
-import com.wq.auth.domain.auth.response.SocialLoginResult
 import com.wq.auth.security.jwt.JwtProvider
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.stereotype.Component
@@ -26,9 +27,10 @@ class GoogleLoginProvider(
         log.info { "Google 소셜 로그인 처리 시작" }
 
         val oauthUser = googleOAuthClient.getUserFromAuthCode(
-            request.authCode,
-            request.codeVerifier,
-            request.redirectUri
+            OAuthAuthCodeRequest(
+                request.authCode,
+                request.codeVerifier
+            )
         )
 
         log.info { "OAuth 사용자 정보 조회 완료: ${oauthUser.email}" }
